pkcs11: Rename config/pkcs11 to config/pkcs11config
